backend/requirements.txt
# Core FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Web3 and Blockchain
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1

# WebSocket and Messaging
websockets==12.0
msgpack==1.0.7

# HTTP and Networking
requests==2.31.0
aiohttp==3.9.1

# Data Processing
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Async and Concurrency
asyncio-mqtt==0.16.1

# Utilities
python-multipart==0.0.6
